Completed all objects.  Two known bugs/errors
Bug 1:
If the truck bin in the mine material handling model doesn't have enough
room for a truck load, a truck in the truck circuit model will never
'endTransfer' on its unload and will become jammed.

Bug2:
Trains/wagons do not wait for the train bin to fill if it is empty.